requirements.txt
streamlit
python-docx
openai
python-dotenv